Add a theme blacklist
There are some theme packages that define multiple themes of which one
is not a valid theme. For example, solarized defines solarized,
solarized dark and solarized light, but the first is not a valid theme
and can't load. Such themes should not be loaded.

An alternative would be to just ignore errors when loading themes and
skip over to the next one if it doesn't work, but I prefer knowing which
themes don't load so I can try to do something about it.